Remove Laplacian pixelation hard-fail from webcam validation.
Daytime Laplacian thresholds were rejecting usable soft scenery (especially Eye-n-Sky push cams). Keep uniform/corrupt/Blue Iris checks; preserve daylight-phase unit coverage in DaylightPhaseTest.
